requirements.txt
absl-py==2.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
billiard==4.2.4
blake3==1.0.8
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.2
cron-descriptor==1.4.5
cryptography==46.0.6
cycler==0.12.1
daphne==4.2.1
Deprecated==1.3.1
dj-database-url==3.1.2
Django==5.2.12
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-encrypted-model-fields==0.6.5
django-timezone-field==7.2.1
djangorestframework==3.17.1
dnspython==2.8.0
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.118.3
fastapi-events==0.12.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
gunicorn==25.3.0
h11==0.16.0
httptools==0.7.1
huggingface_hub==0.36.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==9.0.0
Incremental==24.11.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
kombu==5.6.2
langdetect==1.0.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
ninja==1.13.0
nltk==3.9.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==26.0
pandas==2.3.3
pillow==12.2.0
prompt_toolkit==3.0.52
protobuf==4.25.9
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==3.0.4
Pygments==2.20.0
PyMySQL==1.1.2
pyOpenSSL==26.0.0
pyparsing==3.3.2
pyreadline3==3.5.4
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2026.1.post1
PyWavelets==1.8.0
PyYAML==6.0.3
redis==7.4.0
regex==2026.4.4
requests==2.33.1
rich==14.3.3
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
service-identity==24.2.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.5
sqlparse==0.5.5
starlette==0.48.0
sympy==1.14.0
tabulate==0.10.0
textblob==0.20.0
threadpoolctl==3.6.0
tomli==2.4.1
tqdm==4.67.3
trampoline==0.1.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.43.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
whitenoise==6.12.0
wrapt==2.1.2
wsproto==1.3.2
zipp==3.23.0
zope.interface==8.2
